use a close() to recycle all resources
