+++ spelling_dict.txt
@@ -261,3 +261,4 @@ zabcdefghijklmnopqrstuvwxyz
 zabcdefrstuvwxyz
 zerofill
 analysing
+impls
