+++ sqle/pkg/driver/adaptor_audit.go
@@ -3,18 +3,14 @@ package driver
 import (
 	"context"
 	"database/sql"
-	_driver "database/sql/driver"
 	"os"
 
 	"github.com/actiontech/sqle/sqle/driver"
 	"github.com/actiontech/sqle/sqle/pkg/params"
-	"github.com/actiontech/sqle/sqle/utils"
 
 	"github.com/hashicorp/go-hclog"
 	goPlugin "github.com/hashicorp/go-plugin"
-	"github.com/percona/go-mysql/query"
 	"github.com/pkg/errors"
-	"vitess.io/vitess/go/vt/sqlparser"
 )
 
 // AuditAdaptor is a wrapper for the sqle driver layer. It
@@ -108,11 +104,16 @@ func (a *AuditAdaptor) GeneratePlugin(opts ...AdaptorOption) goPlugin.Plugin {
 	}
 
 	newDriver := func(cfg *driver.Config) driver.Driver {
+		if p, exist := pluginImpls[driver.PluginNameAuditDriver]; exist {
+			return p
+		}
+
 		a.cfg = cfg
 
 		di := &pluginImpl{a: a}
 
 		if cfg.DSN == nil {
+			pluginImpls[driver.PluginNameAuditDriver] = di
 			return di
 		}
 
@@ -131,6 +132,7 @@ func (a *AuditAdaptor) GeneratePlugin(opts ...AdaptorOption) goPlugin.Plugin {
 
 		di.db = db
 		di.conn = conn
+		pluginImpls[driver.PluginNameAuditDriver] = di
 		return di
 	}
 
@@ -184,164 +186,4 @@ func (r *auditRegistererImpl) Rules() []*driver.Rule {
 
 func (r *auditRegistererImpl) AdditionalParams() params.Params {
 	return r.additionalParams
-}
-
-type auditDriverImpl struct {
-	a    *AuditAdaptor
-	db   *sql.DB
-	conn *sql.Conn
-}
-
-func (d *auditDriverImpl) Close(ctx context.Context) {
-	if err := d.conn.Close(); err != nil {
-		d.a.l.Error("failed to close connection in driver adaptor", "err", err)
-	}
-	if err := d.db.Close(); err != nil {
-		d.a.l.Error("failed to close database in driver adaptor", "err", err)
-	}
-}
-
-func (d *auditDriverImpl) Ping(ctx context.Context) error {
-	if err := d.conn.PingContext(ctx); err != nil {
-		return errors.Wrap(err, "ping in driver adaptor")
-	}
-	return nil
-}
-
-func (d *auditDriverImpl) Exec(ctx context.Context, sql string) (_driver.Result, error) {
-	res, err := d.conn.ExecContext(ctx, sql)
-	if err != nil {
-		return nil, errors.Wrap(err, "exec sql in driver adaptor")
-	}
-	return res, nil
-}
-
-func (d *auditDriverImpl) Tx(ctx context.Context, sqls ...string) ([]_driver.Result, error) {
-	var (
-		err error
-		tx  *sql.Tx
-	)
-
-	tx, err = d.conn.BeginTx(ctx, nil)
-	if err != nil {
-		return nil, errors.Wrap(err, "begin tx in driver adaptor")
-	}
-
-	defer func() {
-		if err != nil {
-			if err = tx.Rollback(); err != nil {
-				err = errors.Wrap(err, "rollback tx in driver adaptor")
-				return
-			}
-		} else {
-			if err = tx.Commit(); err != nil {
-				err = errors.Wrap(err, "commit tx in driver adaptor")
-				return
-			}
-		}
-	}()
-
-	results := make([]_driver.Result, 0, len(sqls))
-	for _, sql := range sqls {
-		result, e := tx.ExecContext(ctx, sql)
-		if e != nil {
-			err = errors.Wrap(e, "exec sql in driver adaptor")
-			return nil, err
-		}
-		results = append(results, result)
-	}
-
-	return results, nil
-}
-
-func (d *auditDriverImpl) Schemas(ctx context.Context) ([]string, error) {
-	rows, err := d.conn.QueryContext(ctx, d.a.dt.ShowDatabaseSQL())
-	if err != nil {
-		return nil, errors.Wrap(err, "query database in driver adaptor")
-	}
-	defer rows.Close()
-
-	var schemas []string
-	for rows.Next() {
-		var schema string
-		if err := rows.Scan(&schema); err != nil {
-			return nil, errors.Wrap(err, "scan database in driver adaptor")
-		}
-		schemas = append(schemas, schema)
-	}
-
-	if rows.Err() != nil {
-		return nil, errors.Wrap(rows.Err(), "scan database in driver adaptor")
-	}
-
-	return schemas, nil
-}
-
-func (d *auditDriverImpl) Parse(ctx context.Context, sql string) ([]driver.Node, error) {
-	sqls, err := sqlparser.SplitStatementToPieces(sql)
-	if err != nil {
-		return nil, errors.Wrap(err, "split sql")
-	}
-	if err != nil {
-		return nil, errors.Wrapf(err, "split sql %s error", sql)
-	}
-
-	nodes := make([]driver.Node, 0, len(sqls))
-	for _, sql := range sqls {
-		n := driver.Node{
-			Text:        sql,
-			Type:        classifySQL(sql),
-			Fingerprint: query.Fingerprint(sql),
-		}
-		nodes = append(nodes, n)
-	}
-	return nodes, nil
-}
-
-func classifySQL(sql string) (sqlType string) {
-	if utils.HasPrefix(sql, "update", false) ||
-		utils.HasPrefix(sql, "insert", false) ||
-		utils.HasPrefix(sql, "delete", false) {
-		return driver.SQLTypeDML
-	}
-
-	return driver.SQLTypeDDL
-}
-
-func (d *auditDriverImpl) Audit(ctx context.Context, sql string) (*driver.AuditResult, error) {
-	var err error
-	var ast interface{}
-	if d.a.ao.sqlParser != nil {
-		ast, err = d.a.ao.sqlParser(sql)
-		if err != nil {
-			return nil, errors.Wrap(err, "parse sql")
-		}
-	}
-
-	result := driver.NewInspectResults()
-	for _, rule := range d.a.cfg.Rules {
-		handler, ok := d.a.ruleToRawHandler[rule.Name]
-		if ok {
-			msg, err := handler(ctx, rule, sql)
-			if err != nil {
-				return nil, errors.Wrapf(err, "audit SQL %s in driver adaptor", sql)
-			}
-			result.Add(rule.Level, msg)
-		} else {
-			handler, ok := d.a.ruleToASTHandler[rule.Name]
-			if ok {
-				msg, err := handler(ctx, rule, ast)
-				if err != nil {
-					return nil, errors.Wrapf(err, "audit SQL %s in driver adaptor", sql)
-				}
-				result.Add(rule.Level, msg)
-			}
-		}
-	}
-
-	return result, nil
-}
-
-func (d *auditDriverImpl) GenRollbackSQL(ctx context.Context, sql string) (string, string, error) {
-	return "", "", nil
 }
